WSP-1855 "pegUrl problem with Firefox"
Problem was that default flash pageUrl works bad with Firefox.
It sends only domain info, and not sends particular page info
(like http://domain.com instead of http://domain.com/page.html)
Now we checked in JS if its Firefox and if so - send pageUrl
from JS layer (window.location). On server side we checked if
pageUrlFromJS != 0 and if so - use it.